pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "big-llm-task"
version = "0.1.0"
requires-python = ">=3.11"
description = "CLI utility for large-context LLM queries using a job-directory convention."
dependencies = [
    "anthropic>=0.40",
    "boto3>=1.35",
    "httpx>=0.27",
    "typer>=0.12",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.scripts]
big-llm-task = "big_llm_task.cli:app"
